# Introduce ArrayView type for simple functions 