--- a/velox/core/CoreTypeSystem.h
+++ b/velox/core/CoreTypeSystem.h
@@ -171,6 +171,7 @@ class ArrayValWriter {
   const_iterator end() const {
     return values_.end();
   }
+
   const_reference at(size_t index) const {
     return values_.at(index);
   }
@@ -191,6 +192,18 @@ class ArrayValReader : public ArrayValWriter<VAL> {
     for (auto& val : vals) {
       append(std::move(val));
     }
+  }
+
+  bool mayHaveNulls() const {
+    return false;
+  }
+
+  std::optional<VAL> operator[](size_t index) const {
+    return {ArrayValWriter<VAL>::at(index)};
+  }
+
+  std::optional<VAL> at(size_t index) const {
+    return {ArrayValWriter<VAL>::at(index)};
   }
 };