--- a/velox/expression/tests/CMakeLists.txt
+++ b/velox/expression/tests/CMakeLists.txt
@@ -13,8 +13,9 @@
 # limitations under the License.
 
 add_executable(
-  velox_expression_test ExprTest.cpp CastExprTest.cpp EvalSimplifiedTest.cpp
-                        SignatureBinderTest.cpp SimpleFunctionTest.cpp)
+  velox_expression_test
+  ExprTest.cpp CastExprTest.cpp EvalSimplifiedTest.cpp SignatureBinderTest.cpp
+  SimpleFunctionTest.cpp ArrayViewTest.cpp)
 
 add_test(
   NAME velox_expression_test