--- a/velox/functions/prestosql/benchmarks/ArrayContainsBenchmark.cpp
+++ b/velox/functions/prestosql/benchmarks/ArrayContainsBenchmark.cpp
@@ -26,12 +26,32 @@ namespace {
 
 template <typename T>
 VELOX_UDF_BEGIN(contains)
-// UDF to tell whether a element is part of an array
 FOLLY_ALWAYS_INLINE bool call(
     bool& out,
-    const arg_type<Array<T>>& x,
-    const arg_type<T>& element) {
-  out = std::find(x.begin(), x.end(), element) != x.end();
+    const arg_type<Array<T>>& array,
+    const arg_type<T>& key) {
+  if (array.mayHaveNulls()) {
+    auto nullFound = false;
+    for (const auto& item : array) {
+      if (item && (*item == key)) {
+        out = true;
+        return true;
+      }
+      nullFound |= !item.has_value();
+    }
+    if (!nullFound) {
+      out = false;
+    }
+    return !nullFound;
+  }
+  // Not nulls path
+  for (const auto& item : array) {
+    if (*item == key) {
+      out = true;
+      return true;
+    }
+  }
+  out = false;
   return true;
 }
 VELOX_UDF_END();
@@ -46,12 +66,6 @@ class ArrayContainsBenchmark : public functions::test::FunctionBenchmarkBase {
         bool,
         facebook::velox::Array<int32_t>,
         int32_t>({"contains_alt"});
-
-    registerFunction<
-        udf_contains<Varchar>,
-        bool,
-        facebook::velox::Array<Varchar>,
-        Varchar>({"contains_alt"});
   }
 
   void runInteger(const std::string& functionName) {
@@ -111,7 +125,7 @@ class ArrayContainsBenchmark : public functions::test::FunctionBenchmarkBase {
   }
 };
 
-BENCHMARK(vectorAdapterInteger) {
+BENCHMARK(vectorSimpleFunction) {
   ArrayContainsBenchmark benchmark;
   benchmark.runInteger("contains_alt");
 }
@@ -119,16 +133,6 @@ BENCHMARK(vectorAdapterInteger) {
 BENCHMARK_RELATIVE(vectorFunctionInteger) {
   ArrayContainsBenchmark benchmark;
   benchmark.runInteger("contains");
-}
-
-BENCHMARK(vectorAdapterVarchar) {
-  ArrayContainsBenchmark benchmark;
-  benchmark.runVarchar("contains_alt");
-}
-
-BENCHMARK_RELATIVE(vectorFunctionVarchar) {
-  ArrayContainsBenchmark benchmark;
-  benchmark.runVarchar("contains");
 }
 
 } // namespace