This is continuous and extension work of 
https://github.com/facebookincubator/velox/pull/493

## ArrayView 
Introduce and use ArrayView<T> as an input type for simple functions for ArrayType. 
ArrayView is a lightweight lazy access abstraction for the underlying array 
with the following interface.
- size()
- operator [ ], at returns an ArrayView<T>::Element

Element is an abstract over the value and nullability
 of the array at a specific index, it have an interface similar to std::optional:
 
- value()  // for now unchecked access, (shall we make it checked?) i.e add a dynamic check that value is not null
- operator *  // unchecked access
- has_value()
-  implicit bool cast => if(v)  == if(v.has_value())

ArrayView provides an iterator interface to  iterate over the elements

i.e:
```
foo( ArrayView<T> inputs){
  for(const auto & element: inputs){
      if(element.has_value()){
         bar(*element); 
       }
  }
}
```
or 
```
foo( ArrayView<T> inputs){
  for(int i=0; i<inputs.size(); i++){
      if(inputs[i].has_value()){
         bar(*inputs[i]); 
       }
  }
}
```

or 
```
foo( ArrayView<T> inputs){
 auto it = inputs.begin(); 
 while(it!=inputs.end()){
    if(*it.has_value()){
             bar(**it); 
    }
    ++it;
 }
}
```

## Performance:
this diff speedup simple functions by 2X, simple functions are not yet as fast as optimized vector functions, but they are almost  as fast as vector functions that do not have special fast paths based on dynamic info not available to simple functions.

### 1) min function
VectorMinIntegerBasic is added and it computes min with out any vector-function privileged optimizations
It is added as a comparison point to observe the overhead (if any) of the argument dispatch in the vector reader.
the performance of vectorMinIntegerBasic and simpleMinIntegerIterator are similar 
```
============================================================================
fastMinInteger                                             488.62us    2.05K
vectorMinInteger                                  84.74%   576.58us    1.73K
vectorMinIntegerNoFastPath                        58.81%   830.82us    1.20K
simpleMinIntegerIterator                          58.75%   831.76us    1.20K
simpleMinInteger                                  56.18%   869.66us    1.15K
============================================================================
```
#### before and after this diff
the simple function min implementation did not support null reading:
so here we compare non-nulls min
before: 1.58ms
after: 803.24us 
### 2) compare  function
- with current vector function 
```
with this diff
============================================================================
vectorSimpleFunction                                         1.04ms   961.28
vectorFunctionInteger                            202.30%   514.24us    1.94K
============================================================================

before this diff
============================================================================
SimpleFunction                                         2.11ms   473.00
vectorFunction                            364.69%   579.71us    1.72K
============================================================================
```
- with the fast path under those conditions disabled in the vector function:
(elementsDecoded.isIdentityMapping() &&
      !elementsDecoded.mayHaveNulls() && searchDecoded.isConstantMapping())
```
with this diff
============================================================================
vectorSimpleFunction                                         1.03ms   974.86
vectorFunctionInteger                            128.00%   801.37us    1.25K
============================================================================

before this diff
============================================================================
SimpleFunction                                         1.91ms   523.35
vectorFunction                            233.55%   818.14us    1.22K
```

 
## Test 
- add ArrayViewTest
- simpleFunctionTest already covers cases of nested complex types.

## Next
- introduce View types for row, map .
- introduce 0-copy write proxies for complex types.