@laithsakka 2021-10-29
support all the iterator_traits for c++ input_iterator