@laithsakka 2021-10-29
> Looks good to me. Curious, what do benchmarks show? How much faster/slower simple "contains" and "array_min/max" are compared to vector ones and hand-written loops.
> 
> CC: @damianr99

1- what current benchmark show:
so the simple max_min performance is the same as the vector function implementation if we exclude the fast path
under those conditions : "if (elementsDecoded.isIdentityMapping() && !elementsDecoded.mayHaveNulls()) "

We can in the future try to specialize vector readers for common cases i.e sIdentityMapping(), no null, flat vectors.
to get something faster.

2. I will add the compare benchmark results to the diff description.