@laithsakka 2021-11-02
- address comments..
- support < for iterator.
- update contain simple implementation to match the vector.
- rebase