@laithsakka 2021-11-02
> Overall looks really good!
> 
> So...how do I use this helper interface in my vector function implementations?

you can create a vector reader the same way its used in the test case, from the decoded results.